set lspreference{text,write,read}
---

https://microsoft.github.io/language-server-protocol/specifications/specification-3-15/

The document highlight request is sent from the client to the server to resolve
a document highlights for a given text document position. For programming
languages this usually highlights all references to the symbol scoped to this
file

---

https://microsoft.github.io/language-server-protocol/specifications/specification-3-17/#documentHighlightKind

Text  = A textual occurrence.
Read  = Read-access of a symbol, like reading a variable.
Write = Write-access of a symbol, like writing to a variable.